+++ electron/main.js
@@ -1,14 +1,12 @@
-const { app, BrowserWindow, ipcMain } = require('electron');
+const { app, BrowserWindow } = require('electron');
 
 const path = require('path');
-const Store = require('electron-store');
 const isDev = require('electron-is-dev');
 
 const { getMainWindow, setMainWindow } = require('./window');
 const { generateMenu } = require('./menu');
-const { GH_AUTH_TOKEN, BACKUP_GIST_ID } = require('./constants');
 
-const store = new Store();
+require('./actions');
 
 const createWindow = () => {
   const mainWindow = new BrowserWindow({
@@ -80,30 +78,4 @@ app.on('activate', () => {
   if (getMainWindow() === null) {
     createWindow();
   }
-});
-
-ipcMain.on('LOAD_GH_AUTH_DATA', (event) => {
-  const token = store.get(GH_AUTH_TOKEN);
-  const backupGistId = store.get(BACKUP_GIST_ID);
-  event.sender.send('LOAD_GH_AUTH_DATA_REPLY', {
-    token,
-    backupGistId
-  });
-});
-
-ipcMain.on('SET_GH_AUTH_DATA', (event, data) => {
-  const { token, backupGistId } = data;
-
-  if (token) {
-    store.set(GH_AUTH_TOKEN, token);
-  }
-
-  if (backupGistId) {
-    store.set(BACKUP_GIST_ID, backupGistId);
-  }
-});
-
-ipcMain.on('DELETE_GH_AUTH_DATA', () => {
-  store.delete(GH_AUTH_TOKEN);
-  store.delete(BACKUP_GIST_ID);
 });
