+++ src/components/SnippetList/AccountModal/AuthTokenPanel.js
@@ -23,8 +23,13 @@ const AuthTokenPanel = ({ authToken, onAuthTokenChange, onAccept, loading }) => 
         </FormGroup>
 
         <div className={Classes.DIALOG_FOOTER_ACTIONS}>
-          {/* <Button disabled={!authToken}>Unlink</Button> */}
-          <Button disabled={!authToken} onClick={onAccept} loading={loading}>Connect</Button>
+          <Button
+            disabled={!authToken}
+            onClick={onAccept}
+            loading={loading}
+          >
+            Connect
+          </Button>
         </div>
       </div>
     </>
